#20,#24 Refactor bot code
Resolves #20,#24
