+++ cmd/version.go
@@ -9,18 +9,13 @@ import (
 	"github.com/spf13/cobra"
 )
 
-var version = "unknown"
+var version = "develop"
 
 // versionCmd represents the version command
 var versionCmd = &cobra.Command{
 	Use:   "version",
-	Short: "A brief description of your command",
-	Long: `A longer description that spans multiple lines and likely contains examples
-and usage of using your command. For example:
-
-Cobra is a CLI library for Go that empowers applications.
-This application is a tool to generate the needed files
-to quickly create a Cobra application.`,
+	Short: "Show build version",
+	Long:  `Show build version`,
 	Run: func(cmd *cobra.Command, args []string) {
 		fmt.Printf("i-will-remind-you-bot version: %s\n", version)
 	},
